requirements.txt
flask==2.0.1
gunicorn==20.0.4
tensorflow==2.5.0
pillow==8.1.0